back updated

+++ server/cinema-app-back/src/main/java/com/cinema/service/impl/JpaUserService.java
@@ -32,7 +32,7 @@ public class JpaUserService implements UserService {
 
     @Override
     public List<Users> findAll() {
-        return userRepository.findAll();
+        return userRepository.findByDeleted(false);
     }
 
     @Override
@@ -47,11 +47,18 @@ public class JpaUserService implements UserService {
     }
 
     @Override
-    public void delete(Long id) {
+    public Users delete(Long id) {
         Optional<Users> userOptional = userRepository.findById(id);
-        if (userOptional.isPresent()) {
+        if (userOptional.isPresent() && userOptional.get().getTickets().isEmpty()) {
             userRepository.deleteById(id);
+            return userOptional.get();
         }
+        if(userOptional.isPresent() && !userOptional.get().getTickets().isEmpty()) {
+        	userOptional.get().setDeleted(true);
+        	return userRepository.save(userOptional.get());
+        }
+        
+      return null;
     }
 
     @Override
@@ -85,4 +92,15 @@ public class JpaUserService implements UserService {
 
         return true;
     }
+
+	@Override
+	public Users changeRole(Users user, String role) {
+		if(role.equals("USER") || role.equals("ADMIN")) {
+			user.setRole(UserRole.valueOf(role));
+			Users changedUser = userRepository.save(user);
+			return changedUser;
+		}
+	
+		return null;
+	}
 }
